# api/openapi.yaml
openapi: 3.0.3

info:
  title: Task Management API
  version: 1.0.0
  description: A simple task management API built with Go and Gin.

servers:
  - url: http://localhost:8080

paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /tasks:
    post:
      operationId: createTask
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: getTasks
      summary: Get all tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  /tasks/{id}:
    get:
      operationId: getTaskByID
      summary: Get task by ID
      parameters:
        - $ref: "#/components/parameters/TaskID"
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateTask
      summary: Update task by ID
      parameters:
        - $ref: "#/components/parameters/TaskID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteTask
      summary: Delete task by ID
      parameters:
        - $ref: "#/components/parameters/TaskID"
      responses:
        "204":
          description: Task deleted successfully
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  parameters:
    TaskID:
      name: id
      in: path
      required: true
      schema:
        type: string

  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok

    Task:
      type: object
      required:
        - id
        - title
        - description
        - completed
      properties:
        id:
          type: string
          example: "1"
        title:
          type: string
          example: Learn Go
        description:
          type: string
          example: Build Task API
        completed:
          type: boolean
          example: false

    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: Learn Go
        description:
          type: string
          example: Build Task API

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          example: Learn Go deeply
        description:
          type: string
          example: Gin project
        completed:
          type: boolean
          example: true

    ErrorResponse:
      type: object
      required:
        - message
        - code
      properties:
        message:
          type: string
          example: task not found
        code:
          type: string
          example: TASK_NOT_FOUND